misc: merge allanime crypto into provider
